refactor(engine): extract spill/buffer.rs (SPL-4 #2326)
